pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Andrei Gheorghe <gheorgheandrei13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pywebview = "^5.1"
torch = "^2.0.1"
demucs = {git = "https://github.com/facebookresearch/demucs"}
numpy = "1.26.4"
basic-pitch = "^0.3.3"
protobuf = "^3.20.0"
tensorflow-macos = ">=2.4.1,<2.15.1"
simpleaudio = "^1.0.4"
flask = "^3.0.3"
types-flask = "^1.1.6"
watchdog = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-sugar = "*"
pytest-cov = "*"
black = "*"
mypy = "*"
ruff = "*"
isort = "*"
tox = "^4.11.4"
requests = "^2.32.3"
types-requests = "^2.32.0.20240712"

[tool.poetry.scripts]
app = "src.main:run"

[tool.pyright]
reportPrivateImportUsage = "none"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
